refactor: copy element
