--- src/lib/components/CopyButton.tsx
@@ -1,52 +1,0 @@
-import { CopyIcon } from "@chakra-ui/icons";
-import { Button, Tooltip, useClipboard } from "@chakra-ui/react";
-import type { ButtonProps } from "@chakra-ui/react";
-import { useEffect } from "react";
-
-import { AmpEvent, AmpTrack } from "lib/services/amplitude";
-
-interface CopyButtonProps {
-  isDisable?: boolean;
-  value: string;
-  size?: ButtonProps["size"];
-  copyLabel?: string;
-}
-
-export const CopyButton = ({
-  isDisable,
-  value,
-  size = "sm",
-  copyLabel = "Copied!",
-}: CopyButtonProps) => {
-  // TODO: revisit useClipboard later
-  const { onCopy, hasCopied, setValue } = useClipboard(value);
-
-  useEffect(() => setValue(value), [value, setValue]);
-
-  return (
-    <Tooltip
-      hasArrow
-      isOpen={hasCopied}
-      label={copyLabel}
-      placement="top"
-      arrowSize={8}
-      bg="honeydew.darker"
-    >
-      <div>
-        <Button
-          isDisabled={isDisable}
-          variant="outline-info"
-          size={size}
-          float="right"
-          onClick={() => {
-            AmpTrack(AmpEvent.USE_COPY_BUTTON);
-            onCopy();
-          }}
-          leftIcon={<CopyIcon boxSize="4" onClick={onCopy} />}
-        >
-          Copy
-        </Button>
-      </div>
-    </Tooltip>
-  );
-};
