+++ src/lib/components/ExplorerLink.tsx
@@ -13,7 +13,7 @@ import { AmpTrackMintscan } from "lib/services/amplitude";
 import { truncate } from "lib/utils";
 
 import { AppLink } from "./AppLink";
-import { Copier } from "./Copier";
+import { Copier } from "./copy";
 
 export type LinkType =
   | AddressReturnType
