+++ src/lib/components/copy/index.ts
@@ -1,0 +1,2 @@
+export * from "./CopyButton";
+export * from "./Copier";
