+++ src/lib/components/modal/CodeSnippet.tsx
@@ -19,7 +19,7 @@ import { useWallet } from "@cosmos-kit/react";
 import AceEditor from "react-ace";
 import { MdCode } from "react-icons/md";
 
-import { CopyButton } from "../CopyButton";
+import { CopyButton } from "lib/components/copy";
 import { CustomTab } from "lib/components/CustomTab";
 import { useLCDEndpoint } from "lib/hooks";
 import { AmpEvent, AmpTrack } from "lib/services/amplitude";
