+++ src/lib/pages/instantiate/component/FailedModal.tsx
@@ -15,7 +15,7 @@ import {
 } from "@chakra-ui/react";
 import { IoIosWarning } from "react-icons/io";
 
-import { CopyButton } from "lib/components/CopyButton";
+import { CopyButton } from "lib/components/copy";
 
 interface FailedModalProps {
   errorLog: string;
