+++ src/lib/pages/query/components/QueryArea.tsx
@@ -7,7 +7,7 @@ import dynamic from "next/dynamic";
 import { useEffect, useState } from "react";
 
 import { ContractCmdButton } from "lib/components/ContractCmdButton";
-import { CopyButton } from "lib/components/CopyButton";
+import { CopyButton } from "lib/components/copy";
 import JsonInput from "lib/components/json/JsonInput";
 import JsonReadOnly from "lib/components/json/JsonReadOnly";
 import { DEFAULT_RPC_ERROR } from "lib/data";
